+++ app/org/fairs/[fairId]/sales/page.tsx
@@ -42,6 +42,8 @@ export default async function OrgSalesFeedPage({
         initialTotalRevenue={initial.totalRevenue}
         initialCashRevenue={initial.cashRevenue}
         initialWalletDonations={initial.walletDonations}
+        initialPoolAssistanceGiven={initial.poolAssistanceGiven}
+        initialStudentsAssisted={initial.studentsAssisted}
         initialPayout={initial.payout}
         initialPayoutIsFinal={initial.payoutIsFinal}
         initialMissingInventoryCost={initial.missingInventoryCost}